Refactored the org class and made a config reader and default config generator
