+++ src/test/java/api_testing/pojo/Coord.java
