+++ src/test/java/api_testing/pojo/Rain.java
