+++ src/test/java/api_testing/pojo/Snow.java
