+++ src/test/java/api_testing/pojo/WeatherItem.java
