+++ src/test/java/api_testing/pojo/Wind.java
